refactor: Extract MCP tools into dedicated classes
Move the inline tool/resource registrations out of McpServer into
dedicated classes under the Artemeon\M2G\Service\Mcp namespace. Each
tool now owns its Tool definition and execute() handler, exposed via
the McpTool/McpResource interfaces. McpServer becomes a thin registry.
